# Conflicts:
#	src/main/java/ar/edu/unlam/tallerweb1/domain/cuidado/Cuidado.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/mascotas/ServicioMascotaImpl.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/ServicioUsuario.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/Usuario.java
#	src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorLoginTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorMascotaTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/domain/ServicioLoginImplTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/domain/ServicioMascotaTest.java

+++ src/main/java/ar/edu/unlam/tallerweb1/delivery/ControladorMascotas.java
@@ -67,7 +67,7 @@ public class ControladorMascotas {
     }
 
     @RequestMapping(path = "/mascotas/usuario")
-    public ModelAndView getMascotaPorIdUsuario(long idUsuario) {
+    public ModelAndView getMascotaPorIdUsuario(int idUsuario) {
         ModelMap model = new ModelMap();
         List<Mascota> result = this.servicioMascota.obtenerMascotaPorIdUsuario(idUsuario);
         model.put("mascotas", result);
@@ -81,7 +81,7 @@ public class ControladorMascotas {
     }
 
     @RequestMapping(path = "/mascotas/mis-mascotas")
-    public ModelAndView getMascotasUsuario(long idUsuario) {
+    public ModelAndView getMascotasUsuario(int idUsuario) {
         ModelMap model = new ModelMap();
         List<Mascota> mascotas = this.servicioMascota.obtenerMascotaPorIdUsuario(idUsuario);
         model.put("mascotas", mascotas);
