+++ src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorMascotaTest.java
@@ -78,8 +78,8 @@ public class ControladorMascotaTest {
     }
     @Test
     public void buscarMascotaPorIdUsuarioIncorrecto(){
-        Long idUsuario = 1l;
-        Long idUsuarioIncorrecto = 2l;
+        int idUsuario = 1;
+        int idUsuarioIncorrecto = 2;
         dadoQueExisteMascota(idUsuario);
         ModelAndView mav = this.controladorMascotas.getMascotaPorIdUsuario(idUsuarioIncorrecto);
         assertThat(mav.getModel().get("mascotas").toString()).isEqualTo("[]");
@@ -213,7 +213,7 @@ public class ControladorMascotaTest {
     }
     @Test
     public void buscarMascotaPorIdUsuario(){
-        Long idUsuario = 1l;
+        int idUsuario = 1;
         dadoQueExisteMascota(idUsuario);
         ModelAndView mav = this.controladorMascotas.getMascotaPorIdUsuario(idUsuario);
         assertThat(mav.getModel().get("mascotas")).isNotNull();
